+++ src/plugins/eip7002/eip7002_plugin.c
@@ -37,7 +37,7 @@ static void eip7002_plugin_init_contract(ethPluginInitContract_t *param) {
     eip7002_context_t *context = (eip7002_context_t *) param->pluginContext;
 
     explicit_bzero(context, sizeof(*context));
-    if ((context->received + CALLDATA_SELECTOR_SIZE) > sizeof(context->withdrawal_request)) {
+    if ((size_t) context->received + CALLDATA_SELECTOR_SIZE > sizeof(context->withdrawal_request)) {
         param->result = ETH_PLUGIN_RESULT_ERROR;
     } else {
         memcpy(&context->withdrawal_request[context->received],
